상점 페이지 옵션형 카테고리 완성

+++ src/main/java/com/spring/domain/ShopProductVO.java
@@ -13,7 +13,7 @@ public class ShopProductVO {
 	private int amount;
 	private String category;
 	private String description;
-	private int gradeavg;
+	private float gradeavg;
 	private int volume;
 	private int shipcost;
 	private String storage;
